# movies/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movie API",
        "description": "Tariq CA1 Movie API",
        "contact": {
            "email": "N00202437@iadt.ie"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/movies": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Display a list of all Movies.",
                "description": "Displays all the movies",
                "operationId": "a7859dc5cd106bbeae5492e22df893ec",
                "responses": {
                    "200": {
                        "description": "Successful operation, Returns a list of Movies in JSON format"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Movies"
                ],
                "summary": "Update a Movie",
                "description": "Stores the movie in the DB",
                "operationId": "updateMovie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Movie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "genre",
                                    "runtime",
                                    "director",
                                    "rating",
                                    "description",
                                    "release_date",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sample Title"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sample Genre"
                                    },
                                    "runtime": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "120"
                                    },
                                    "director": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sample Director"
                                    },
                                    "rating": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "7.2"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "A long description about this great movie"
                                    },
                                    "release_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-05-05"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "https://picsum.photos/200/300"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Movies"
                ],
                "summary": "Create a new Movie",
                "description": "Stores the movie in the DB",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "genre",
                                    "runtime",
                                    "director",
                                    "rating",
                                    "description",
                                    "release_date",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sample Title"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sample Genre"
                                    },
                                    "runtime": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "120"
                                    },
                                    "director": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sample Director"
                                    },
                                    "rating": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "7.2"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "A long description about this great movie"
                                    },
                                    "release_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-05-05"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "https://picsum.photos/200/300"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/movies/{id}": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Display 1 movie by the id.",
                "description": "Gets a movie by ID",
                "operationId": "a7031f1ec2eb856acbdd7f7054ed001d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Movie id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Movies"
                ],
                "summary": "Delete a Movie",
                "description": "Delete Movie",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a Movie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "204"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}